feature/add-workspaces-back (#41)
Added Workspaces back in order to debug deployments and determine if vercel can support workspaces.
Disabled swcMinify, because it caused vercel build errors (it falls back to terser)
We can revert this PR if it doesn't work out. But we would prefer to try and get workspaces functioning correctly, and vercel have said workspaces are supported.
Added a <Text> component to properly determine if component-library styles are imported from React imports.