fix/HWPD-3538/optional-fields-with-regex
Theer was a bug where if a field was optional, but had a regex, and was hidden from view due to other conditional fields, the validation flagged it as required.
This change adjusts the regex validation logic to accomodate case where regex is the only validation and the field is optional.